\name{prepareLegend}
\alias{prepareLegend}
\title{prepareLegend}
\description{Function create legend image with colored sheres and labels corresponding to the elements colorDef data.frame   }
\usage{prepareLegend(colorDef, windowRect = c(0, 0, 900, 700))}
\arguments{
  \item{colorDef}{Dictionary defining color of the voxel in dependence of its content.}
  \item{windowRect}{position and size of RGL window}
}



\author{Anatoly Sorokin, Oksana Sorokina}





