\name{showRGL}
\alias{showRGL}
\title{showRGL}
\description{Visualisation function. Wrapper to RGL call.  }
\usage{showRGL(res, col = res$Col, alpha = res$Alph, windowRect = c(0, 
    0, 900, 700))}
\arguments{
  \item{res}{structure to be visualized}
  \item{col}{vector of colors for voxels}
  \item{alpha}{vector of  transparancy values for voxels}
  \item{windowRect}{position and size of RGL window}
}



\author{Anatoly Sorokin, Oksana Sorokina}





