\name{SPKappaR-package}
\alias{SPKappaR-package}
\docType{package}
\title{R companion to SPKappa}
\description{Parse output of SPKappa, process and visualize results}
\details{
\tabular{ll}{Package: \tab SPKappaR\cr
Type: \tab Package\cr
Title: \tab R companion to SPKappa\cr
Version: \tab 0.0-0-0\cr
Date: \tab 2012-12-01\cr
Author: \tab Anatoly Sorokin, Oksana Sorokina\cr
Maintainer: \tab Anatoly Sorokin <lptolik@gmail.com>\cr
License: \tab GPL (>= 2)\cr
Copyright: \tab Anatoly Sorokin, Oksana Sorokina 2012\cr
Depends: \tab rgl\cr
Suggests: \tab \cr
Imports: \tab rkappa,RColorBrewer\cr
LazyData: \tab true\cr}
}
\author{Anatoly Sorokin, Oksana Sorokina}

\keyword{ package }


